+++ apps/pizza_api/src/pizza_api/settings/__init__.py
@@ -2,8 +2,10 @@
 
 from pizza_api.settings.app import ApplicationSettings
 from pizza_api.settings.engine_pool import ApiPoolSettings
+from pizza_api.settings.scaleway import ScalewaySettings
 
 app_settings = ApplicationSettings()
 pool = ApiPoolSettings()
+scw = ScalewaySettings()
 
-__all__ = ["app_settings", "pool"]
+__all__ = ["app_settings", "pool", "scw"]